// app/data/pipeline/types.ts
export interface RawAbility {
    id: string;
    name: string;
    legend: string;
    factionId: string;
    description: string;
    type: "Core" | "Faction" | "Datasheet";
    parameter?: string;
}